*Added user permissions to api

+++ project/blog/api/views.py
@@ -1,6 +1,10 @@
 from rest_framework import status
 from rest_framework.response import Response
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.pagination import PageNumberPagination
+from rest_framework.generics import ListAPIView
+from rest_framework.authentication import TokenAuthentication
 
 from django.shortcuts import get_object_or_404
 from ..models import Post
@@ -11,6 +15,7 @@ from django.contrib.auth.models import User
 # region Post detail
 
 @api_view(['GET',])
+@permission_classes((IsAuthenticated,))
 def apiPostDetail(request,year,month,day,post):
     post=get_object_or_404(Post,
                         slug=post,
@@ -28,15 +33,12 @@ def apiPostDetail(request,year,month,day,post):
 
 # region Post list
 
-@api_view(['GET',])
-def apiPostList(request):
-    try:
-        post=Post.objects.all()
-    except:
-        return Response(status=status.HTTP_404_NOT_FOUND)
-    if request.method == 'GET':
-        serializer = PostSerializer(post,many=True)
-        return Response(serializer.data)
+class apiPostList(ListAPIView):
+    queryset = Post.objects.all()
+    serializer_class = PostSerializer
+    authentication_classes = (TokenAuthentication,)
+    permission_classes = (IsAuthenticated,)
+    pagination_class = PageNumberPagination
 
 # endregion
 
@@ -45,6 +47,7 @@ def apiPostList(request):
 # region Post Update
 
 @api_view(['PUT',])
+@permission_classes((IsAuthenticated,))
 def apiPostDetailUpdate(request,year,month,day,post):
     post=get_object_or_404(Post,
                         slug=post,
@@ -52,6 +55,10 @@ def apiPostDetailUpdate(request,year,month,day,post):
                         publish__year=year,
                         publish__month=month,
                         publish__day=day,)
+    user = request.user
+    if post.author != user:
+        return Response({"Response":"Not allowed"})
+    
     if request.method == 'PUT':
         serializer = PostSerializer(post,data=request.data)
         data = {}
@@ -68,6 +75,7 @@ def apiPostDetailUpdate(request,year,month,day,post):
 # region Post delete
 
 @api_view(['DELETE',])
+@permission_classes((IsAuthenticated,))
 def apiPostDetailDelete(request,year,month,day,post):
     post=get_object_or_404(Post,
                         slug=post,
@@ -75,6 +83,10 @@ def apiPostDetailDelete(request,year,month,day,post):
                         publish__year=year,
                         publish__month=month,
                         publish__day=day,)
+    
+    user = request.user
+    if post.author != user:
+        return Response({"Response":"Not allowed"})
     if request.method == 'DELETE':
         operation = post.delete()
         data = {}
@@ -91,10 +103,9 @@ def apiPostDetailDelete(request,year,month,day,post):
 # region Post Create
 
 @api_view(['POST',])
+@permission_classes((IsAuthenticated,))
 def apiPostDetailCreate(request):
-    user = User.objects.get(pk=1)
-
-    post=Post(author=user)
+    post=Post(author=request.user)
 
     if request.method == "POST":
         serializer = PostSerializer(post,data=request.data)
@@ -103,4 +114,5 @@ def apiPostDetailCreate(request):
             return Response(serializer.data)
         return Response(serializer.errors,status=status.HTTP_404_BAD_REQUEST)
 # endregion
+
